Youtube Transcriptor/requirements.txt
youtube_transcript_api
streamlit
google_generativeai
python-dotenv
pathlib